Create Domain-Data-Scrape
This code below works and produces the following below:
(Runs the functionality to pull text from url/website/domain pages). 
I recommend running a ping test first with the system terminal. 
Using the company (RBTC) laptop, I had to disconnect from hardwire, from company wifi, from Cisco, and connect to a hotspot or so, then proceed with the following. 
Please note this code does not work on formatting the information to a table, will work on later.