Create MS-Excel-Address-Standardization
This code below works and produces the following below:
(Standardization for addresses, which can store and change. (Ex: Ave to Avenue) This is plug and play). I asked a coworker if they had anu fun projects that I could work on . This coworker was unaware how to complete this task on a macro and fast level. This allowed me to compare (2) Microsoft Excel files (The customer file) and (SAP Customer Master Data Export), this standardization found 200+ additional matches more than the direct head-to-head Vlookup formula.